az2keyboard/requirements.txt
azure-cosmos==3.1.0
azure-functions==1.0.0b4
azure-functions-worker==1.0.0b8
certifi==2019.6.16
chardet==3.0.4
grpcio==1.20.1
grpcio-tools==1.20.1
idna==2.8
protobuf==3.8.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
yapf==0.27.0
